Accept new / unknown API fields gracefully (#131)
* Accept new / unknown API fields gracefully

From discussions in Montreal, it seems that new fields in the Juju
schema definitions are expected to be handled gracefully even without a
bump to the facade version. That is, facade versions will only be bumped
if a field is removed or changed.

Fixes #130

* Update schema and generated code for juju-2.2-rc1

The agent-version field that was added to ModelInfo is of a type that
requires some special parsing and serialization logic.  This type was
used before but not in an API that we were apparently exercising.