Improve handling of closed connections (#148)
* Prevent circular reference between Monitor and Connection

* Improve monitor status logic

* Stop watcher task cleanly on disconnect

Instead of the watcher task blowing up with an exception when the
connection is closed, it should exit cleanly and allow the Monitor to
report the connection error.

Fixes #147

* Raise ConnectionClosed exception in Model.block_until

Rather than blocking indefinitely when waiting on a model change that
will never happen if the connection gets closed out from under it, this
makes `Model.block_until` raise a `websockets.ConnectionClosed` error so
that it can be caught and dealt with by the client.

* Automatically reconnect lost websocket connections

The receiver or all_watcher should immediately reconnect a lost
connection, or it will be reconnected automatically upon issuance of the
next RPC call.  However, there is still a small chance that the
disconnect could happen between sending a API call and receiving the
response, in which case a websockets.ConnectionClosed error will still
be raised to the caller.  These should be quite rare, though.

* Gracefully handle add_signal_handler failing in a thread in loop.run

* Restart AllWatcher if controller stops it

Fixes conjure-up/conjure-up#965

* Explicitly let the controller know we're stopping the watcher

* Skip reconnect if close was requested