all: use pyrfc3339 instead of dateutil
The pyrfc3339.parse function can be considerably more lightweight
than dateutil.parser.parse_time as it knows the format that
it's expected to parse (and in all the cases we use it, the API
specifically uses RFC3339 format).

Also, the pyrfc3339 package provides an easy way to format
RFC3339 times, which dateutil does not (and we need that for
the next PR which writes Go-style cookie jars).

Furthermore, pyrfc3339 is a dependency already used by
py-macaroon-bakery, so we can share it when we import
that.