implement macaroon discharge
This isn't complete - we're still not writing cookies acquired
during the authentication process back to the cookie jars,
but we'll add that in another PR.

The integration tests rely on the fact that an agent user
has been added to the staging identity manager.
Its key is included in the commit, because it doesn't
grant any rights that aren't easily acquired elsewhere.