refactor connections prior to bakery authentication (#187)
We do a bunch of refactoring in advance of actually adding
the bakery authentication. No additional functionality
as yet (hence no new tests) but this is a breaking API
change in a few ways.

API changes
----

Model and Controller now use the connect method for both connecting to a named model and for connecting to the current model (if the name is empty or None)

Connection.connect is now a class method that returns a new connected connection.

Various fields in Connection, Model and Controller that are really internal-only are given a leading underscore.

Controller and Model now have connection methods not properties, and they raise an exception if there is no current connection.

Other changes that should not break the API
----

Add a bakery_client argument to Connection.connect so that the client can have control over what interaction methods are used and provide a cookie jar.

Use "-q -B" options to scp so that it doesn't print spurious output when it's used.

Avoid unnecessary explicit inheritance from object.

Add is_connected method to Controller and Model.

Made new Connector class factoring out code in common between Controller and Model.

Factor Connection._connect so that we can do the bakery authentication.

Connection has a new connect_params method that returns parameters suitable for passing to Connection.connect.

Factored out commonality from Connection.pinger and Connection.receiver to _Task, making it easier to disable or mock those tasks in tests.