Fix race condition in adding relations
Using Model._wait_for_new without an entity_id can block indefinitely if
the object is already in the model.  This could happen with relations
because the AddRelation API call might not return execution until after
the _watcher task picks up the delta which is before the call to
Model._wait_for_new.  Unfortunately, the API result doesn't return the
relation ID, so we just have to watch the model for the relation with
the right key.

Fixes #191