JujuData abstract base class
This means we won't have to mock the JujuData methods
in order to be able to connect with different credentials
(for example when connecting with macaroon authentication).

It would be nice if the base class method semantics weren't
tied so closely to the underlying data format, and instead
reflected the (very few) operations that we actually need,
which would make it easier to modify its behaviour, but
this should be OK to begin with.