Try connections in parallel and use first to connect
The controller may return endpoints that won't succeed (as they may be
cloud internal addresses) along with one(s) that will.  These can take
some time to fail, so we should be trying all endpoints at the same time
and using the first to successfully connect.

This also prevents the always failing endpoints from masking an error
connecting to the one that should succeed.