Manually create network bridge and wait for addresses
Both Juju stable and edge will automatically create the network bridge
if it doesn't exist during bootstrap, but it seems that there's an issue
in edge such that Juju doesn't wait properly for the addresses to become
available.  Thus, edge fails to bootstrap while stable doesn't.  By
manually creating the network bridge and waiting for the addresses
ourselves, we should avoid that.