Always send auth-tag even with macaroon auth (#217)
* Always send auth-tag even with macaroon auth

Fixes #216

* Add testenv for more easily running examples

* Add test for macaroon auth

Adds a separate testenv for tests that can't be run in parallel with
other tests.  The macaroon auth test is the first, because it modifies
the login type of the default user and thus could break other tests
logging in in the middle of it.

* Fix test for change in discharge response from api.staging.jujucharms.com

* Manually create network bridge and wait for addresses

Both Juju stable and edge will automatically create the network bridge
if it doesn't exist during bootstrap, but it seems that there's an issue
in edge such that Juju doesn't wait properly for the addresses to become
available.  Thus, edge fails to bootstrap while stable doesn't.  By
manually creating the network bridge and waiting for the addresses
ourselves, we should avoid that.

* Fix capitalization error

* Fix workaround to get LXD addresses listening