Fix JSON serialization error for bundle with lxd to unit placement
When a bundle contains lxd placement to a unit name, the full unit
object model instance was ending up getting passed directly to the API.
This change handles that case by resolving it to the proper machine
number instead.

This also has a fix for the workaround version check sometimes failing
if triggered before Model.get_status() had been called.