Implement model.get_action_output (#242)
* Implement model.get_action_output

* Check explicitly for Noneness

* Wait for the action to complete

Wait for action to complete. The excisting code in juju.model for
waiting does not seem to support specifying a timeout so update
docstring to explain that 'wait' is ignored.

* Stop using wait_for_action when for action output

Calling get_action_output should be safe if the action has completed or is
currently in progress. wait_for_action assumes the action is in progress and
will timeout if the action had completed before it was called.