Add ability to mitigate races when using Observers
If an Observer is invoked after issuing a command which mutates the
state of the model then a race condition is created due to the fact
that observers act on model deltas and if the command completes
before the Observer is initialiased then no delta will be generated
for the observer to act upon.

This change attempts to deal with that situation by causing the
watcher to evaluate a fallback function when a timeout is reached.