Purge lxd to prevent unix socket conflicts
The test_add_manual_machine_ssh test seems to be consistently failing on
Travis with a permission error connecting to the LXD unix socket.  After
some investigation, it seems like this might be a conflict with the apt
version of LXD leaving its socket around or somehow getting reinstalled.
This commit tries to ensure that the socket is cleaned up and debug if
lxd is somehow getting installed again.