Handle RedirectError payloads returned by Login RPCs (#303)
updates to support handling of RedirectError payloads

(introduced by juju/juju#10069)

So far we have two specific use-cases for redirects:

JAAS-like redirects (Login fails with CodeRedirect and no payload; client makes an extra RPC to get the redirect details) which the client library should automatically follow.

Login RPCs fail with CodeRedirect and contain details about the controller the model migrated to (hence no extra RPC is required). For UX purposes (see linked PR for the reasoning behind this) we want to show an error to the user instead of transparently redirecting them to the new controller.

To ensure that both cases work as described above, the JujuRedirectException type is augmented with a follow_redirect attribute that is set to true for JAAS-like redirects. The dialing code has been updated to intercept redirect exceptions and either re-raise them or auto-connect to the new controller depending on the follow_redirect value.