Disable hostname checking on controller connection (#305)
The hostnames listed in the controller cert don't include the IP address
which is typically what the endpoint address uses for self-bootstrapped
controllers.  This means the hostname check will always fail for that
case.  However, since we pre-share and trust the cert from the
connection info, along with the endpoint address, it is safe to skip
that check during the connection.

Fixes #302