Schemagen is no longer required
Schemagen is no longer required because Juju itself has incorporated
the schema generation into the workflow of Juju. This should make it
a lot easier to build upon pylibjuju if it's always up to date with
Juju.

Work to do this has already been completed, see:

 1. https://github.com/juju/juju/pull/10336
 2. https://github.com/juju/juju/pull/10338

This should support 2.6 and onwards. Older releases can be done
the old way, but it requires you to have schemagen installed in the
old positions. The error messages might be more cryptic because of
that.