Client changes for backwards compatibility
In order to use an older client with an older api facade, previous
code in the shim code would have to be changed. Instead we progress
this even further by doing two things...

  1. Make all facade methods take an additional kwargs parameter.
  2. Make all parameters in facade methods positional arguments

The changes to one mean that any additional arguments that you want
to pass to newer facades will now pass the facade method signatures
whilst ignored from the construction of the RPC params dict. Plainly
put, anything that isn't named in the method arguments are just
ignored for older clients that don't understand the arg, but will be
picked up by new clients that do.

To make this work in a future compatible way, named positional
arguments are used throughout to ensure that we can in the future
move to just using kwargs.

The ONE caveat here is that we don't expose the clients outside of
the shim, if we do, then the changes we make to any of the facades
will break all downstream code. If that's not the case, we can change
the facade code at will to ensure that it works for us in the best
sense of the word.