Pin facade versions
The following is a very fundamental change in how pylibjuju will
work in the future. Instead of saying we'll work with the latest
and you get everything for free, instead we'll state that these
are the facades we'll work against and you have to put in work to
get the latest and greatest.

The consequences are that you'll have to understand the facade
versions from juju to know how to bump them correctly.