Fix issue #335: Allow provisioning with non-"ubuntu" user
This patch fixes an issue with the SSH provisioner, which previously
always assumed there was an "ubuntu" user with sudo access available on
the target machine.

The provisioner will now honor the username passed to
`model.add_machine`.

Added an additional integration test for testing the root user, as well
as modifying the existing integration test to use a randomized username
w/sudo access.