Hide method default values behind a flag
The following hides the new default values behind a feature flag,
so that experimental features don't impact existing setups.
Unfortunately default values introduce a problem where default
values are useful except when they're not. In this case they're not
valuable when setting default values for instance types. Moving to
the None type elevates this, but it means that we loose some of the
type goodness. We should instead look at where default types are
useful and when they're not and come up with a mechanism for
generating schema that knows when to use them and when not.
Potentially a white and black list?