Handle interface{} params from api server
The following commit changes how we handle `interface{}` from the
api server. Previously any `interface{}` was handled as a
`map[string]interface{}`, but that's not actually the case. In
reality the value returned could be anything. To support this change
we should not validate the value and just ignore it for the user to
handle correctly.