Bundle changes
The following adds a set of bundle changes classes as a way to
abstract the possible issue where breakages upstream cause chaos
downstream.

This is an idea and the real fix is to make sure that juju itself
correctly handles those changes.

The benefit of doing it this way is that we have similar set of
abstractions to the golang source, making it more familiar to all.