Run implementation tests for type classes
The following adds some very basic unit tests for the run implementations
on the type classes. This should help find any issues that are currently
in the bundle handler. These tests aren't 100% coverage, but more in the
happy path range, this is just a start and more should be added as we
go.