Wire up upgrade charm via bundle
The following is a very basic idea on how we can attempt to wire
up a bundle changes for upgrade charm. I've not tested it yet, as
we may run into issues around the AddCharm path. In theory that isn't
required for upgradeCharm as SetCharm should do that for us.