Surpress KeyError when applications are missing
The following surpresses a KeyError when a application or a remote
application is missing. Instead we should return None and correctly
handle that a application or a remote-application could be missing.

This should be safe to do so and the fact that an update from the
all watcher should fill in the gaps eventually. If this is not the
case, we should in the future ensure that a remote-application can't
raise index/key errors when data is missing.

The missing data when raised/thrown criples the library from being
useful, instead we should really handle all the cases these exist and
correctly insert None or a fallback.