Raise error
After having a long internal discussion around this, it's evident that
this will cause errors down the line at some point. Instead we've
decided that it's better to throw a custom error, one that can be caught
individually to force the user of the library to disconnect and
reconnect from the model, to pick up any missing deltas.

This can be done in a way to the examples/model.py where retries can be
forced to happen when an exception is raised.

This code to change this is very rudimental and just exposes a new juju
error - JujuEntityNotFoundError. I've also updated the test to fully
expose the exception at a unit level.