Handle relation matches
The following code handles relation matches with a different approach.
Instead of attempting to find the underlying model application and
shadow applications (remoteApplication, applicationOffer) for an
endpoint, we should use just the application. The matches code can then
validate that the existance of an application exists in the state whilst
it's traversing.

Hopefully cleaning up the role of the endpoint.application property
method, will make things much clearer.

Additionally I've exposed the endpoint application_name, so we can use
that whilst testing if the application name exists.

While this does clean up the code somewhat, I'm still unsure why we
validate that the application exists, other than we can, not whether we
actually should?