Facade 2.7.0 version bumping
The following commit changes the version of pylibjuju to focus on 2.7.0
as the tracking branch. In doing so, it requires bringing in the 2.7.0
schemas and updating the library to Juju 2.7.0.

Unfortunately we can't use python 3.7.x as there is an issue in
atttempting to introspect the types. It seems like there is a breakage
between 3.6.x and 3.7.x and some significant changes are required in
juju/client/facade (ones I don't fully understand atm) before we can use
that.

The client libraries where generated using a lxc container and
installing 3.6.x and using that as a way to bypass the error.

I've created a tracking issue[1] to resolve it.

 1. https://github.com/juju/python-libjuju/issues/377