Model connection failure
The following fixes the issue where a new facade was introduced in 2.7,
but wasn't added to the library. Pylibjuju expects all new facades to
registered so it can negotiate how to speak to Juju. If a new facade is
introduced in to Juju and pylibjuju isn't aware of it, then
unfortunately it can't understand the whole picture.

The changes ensure that we're using the correct facade versions and that
we re-enable a the skipped test. The intermittent test is better to be
intermittent than to release a library release that's broken.

Unfortunately not all integration tests are run and there could be other
issues like that.

---

Additionally I noticed that a model delta type from the all watcher
isn't being handled, we should fix this in the future. For now it seems
safe enough to ignore it.

---

Along with the changes I've updated warnings about deprecated code.