Derived application status
The following updates pylibjuju to the same logic as 2.8.1, whereby the
derived application status that's in juju is ported across to libjuju.

It uses the same logic as juju[1], which can be found here. The code is
really simple and should work as intended to find out what the correct
application status is, if the application status has been set to
"unset".

1. https://github.com/juju/juju/commit/8bc76070c55c5991c0863e8f8c42df4780acdcbe#diff-7f5882040ea646fab639d5c6351b19f2R129