Derive the application status
With changes to juju, the application status was changed for the CLI to
use the model cache to derive the status of the application from the
unit status. Unfortunately the logic didn't reach the all watcher and in
doing so always reports the last application status.

Unfortunately the facade version wasn't bumped, which makes facade
checking for old and new logic almost impossible. Instead the solution
here is to always query the unit status, BUT add the application status
in to the derive_status() call so that we always buble the most relavant
serverity to the top.

This should happily work in a backwards compatibility way, so if the
application is blocked and any units are waiting, then the application
status will always win. Similarly if the units are blocked and the
application status is waiting, then blocked will be shown.

The one caveat is, that if the older juju didn't bubble up the worst
serverity to the application status, then you could end up with a
missmatch. For example; unit sets to error, but the application status
is waiting. Then juju CLI would show waiting, but pylibjuju would show
error.

I'm unsure if that is actually a fundamental problem, as I don't believe
it should happen, but in theory it should be at least known.