Handle strict mode
Strict mode is a new feature in Juju 2.9. The mode allows more things to
fail early if the mode is switched to strict. If no mode is detected
then it will fallback to a more passive mode, where it will fallback to
other solutions and try to carry on.

The purpose of strict mode is to enable people developing with
Juju/pylibjuju is highlight errors more early on and to prevent less
magic in Juju.

The PR for implementing strict mode is simple, EXCEPT when I turned it
one, it found the first error! ModelInfo delta was wrongly implemented
and wasn't gaining all the history correctly from the deltas over time.

This requires[1] juju/juju#12183 to land before this PR is really
useful.

To enable it at bootstrap time:

```sh
juju bootstrap lxd test --config mode=strict
```

 1. https://github.com/juju/juju/pull/12183