Type name override
Improve the name of the method that allows us to override the type name.
Offically we would like to call ModelInfo - Model, but Model already
exists and I don't want Model to also become a type delta. Model is
both a entity for state and an entity for connections and to become an
entity for deltas seems like a bit too much.