Upgrade trusted to use new facade method
Application facade has a new method for calling configuration values on
it. Use that instead of the old way. This is a breaking change if you're
using it from pylibjuju and we should probably be aware of that in the
future.

Luckily the fix is to just use the latest pylibjuju with this fix in it.