CharmHub Info
The following introduces queries to the charmhub store API. Now that all
requests go through a juju controller it makes it very trivial for other
clients like pylibjuju to gain the same powers as the go CLI.

The code is rather simple, it farms off a new type to encapsulate the
charmhub related calls (info and find which will follow) as the model
type is large and unwieldly, once there we just create a new facade and
perform the query.

Offically it would be nice to have to channel validation, but for now
we'll let the controller handle that.