Improve readability by extracting code
Only looking over the code did I see that there already exists a charm
store and so we should make it easier for others reading this to see
that this exists along with charm hub. Extracting the contents into
another file shows this.

This is a mechanical change that helps with readability, nothing else.