Fix bundles referencing .charm files and add wait_for_bundle
Bundles which referenced local `.charm` files were failing.

This also adds a `Model.wait_for_bundle()` helper, which waits for all
of the units deployed by the bundle to come up and for all hooks to
settle. It is not as powerful as `juju-wait`, but it at least provides
an easy way to block until a deployed bundle has settled into a steady
state.