Make Model.wait_for_idle more generic
Split `wait_for_bundle` off into `utils` and made the helper on `Model`
more generic. Also removed an errant `print` with an f-string and
updated the GitHub Workflow to run lint and unit tests against all
versions to ensure that things like that get caught quickly (as it broke
on Python 3.5).