Remove machine work around
The following change removes the workaround for the 2.2.3 release and
subsequent 2.3-beta1 release. As these are older releases that users
should have moved away from, it's time to remove the workarounds.

The code in question was highly contentious anyway, as it was never
awaited on the machine, so might not have been doing what we wanted it
to do in the long run.