Avoid directly interacting with the websocket object within connection.py
The websocket object (_ws) is tightly coupled with the asyncio-managed
tasks (e.g. Pinger) as well as the Monitor, so interacting with it
directly (e.g. doing conn._ws.close() instead of conn.close()) pushes
things out of sync/control.

This also changes the field name "ws" into "_ws" to indicate that the
field shouldn't be accessed directly.

Some tests are left using the _ws object explicitly because the test
involved is particular to the behavior when _ws object is being used
directly.