Ensure temporarily spawned Controllers are disconnected after use
Functions such as model.get_controller returns a newly constructed
Controller object that is already connected with the model's own
connection parameters. This puts the responsibility of closing out the
connection (and destroying the spawned asyncio tasks etc) to the
caller. If not done, everyone (asyncio, websockets etc) yells at
pylibjuju about lingering tasks and open connections.

Here we introduce a ConnectedController object that when it's used
with the "async with" form automatically disconnects after done.