Introduce create_task_with_handler wrapper to handle exceptions
asyncio loop event_handler is only called on Task exceptions when the
Task object is cleared from the memory. But the garbabe collector
doesn't clear the Task if we keep a reference for it (e.g. putting it
in a neverending Task ---like Pinger) until the very end.

Installing our own handler makes sure that the exceptions are
retrieved and properly handled and logged regardless of when the Task
is being destroyed.