Ensure all watchers validate for the Id
Currently all watchers in Juju are expected to ensure that they pass
back an id. This is currently done per watcher, what should actually
happen is we should correctly wire up the calls to the facades, but this
doesn't seem easy. Instead, this approach ensures that all watchers when
doing an RPC call have a Id. If no Id is present raise an error so that
the correct watcher id can be passed in.