Unit public address
In order to get the correct public address from the unit, you have to do
a query to the application facade to attempt to retrieve it. In order to
support backwards compatibility with older instances of Juju that didn't
have the new facade method, we fall back to the previous implementation.