Use public-address key instead of public_address
PR #600 updated the get_public_address method for units to look up the
public address in the ApplicationFacade's UnitsInfo object. However,
the change used the get(str) method of the Type class which performs a
lookup in the json dictionary using the hyphenated key rather than the
underscore key, thus always returns None since this key is not in the
JSON dictionary. Changing it to return the lookup of 'public-address'
instead of 'public_address' results in the unit's public address being
returned.

Fixes #551

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>