Fix charm resolution for Juju 2.8.11
Fixes #623

This changes two things:

1 - It patches the deployment of CharmStore charms

2 - Normally pylibjuju assumes the CharmHub when it receives a charm
url without a prefix. This change adds an override for the case where
the CharmsFacade version is too low to resolve the charm.

_resolve_charm literally has a failing check for this because lower
versions (such as Juju 2.8) cannot deploy from the CharmHub in the
first place, so it's safe to put in this override and try to get the
charm from the CharmStore instead of failing.