Ensure num_units is adjusted correctly for subordinate charms
Subordinate charms are by definition don't require additional
units. So num-units should be 0 at deploy.

Normally in libjuju num_units defaults to 1, this change lets libjuju
to check the metadata from charmhub to see if we're deploying a
subordinate charm. If that's the case then we adjust the num-units to
0 to be able to deploy it without any problems from Juju's side.

fixes #639