Increase timeout limit for test case
This test seems to be intermittently failing on CI with a timeout, but
despite the frequency of failure, there doesn't seem to be any
fundamental reason for it to fail and it works locally %100 of
time. So trying to increase the timeout limit to see if all it needs
is more time.