Avoid ignoring asyncio exceptions in coroutines
Exceptions that happen when a coroutine is being destroyed are
ignored, but they cause backround noise such as "Exception ignored in
coroutine ..." exceptions etc. This change tries to propogate the
exceptions so that the asyncio base handler itself can ignore the
exception as an attempt to get rid of that exception ignored noise in
test outputs etc.

Should fix #657