Pass series info into origin for ResolveCharm
Without the series the ResolveCharm will select the latest base for
the charm. E.g. even if we want focal, the ResolveCharm will return
22.04 (jammy) for the base channel in the resulting origin if the
charm supports both jammy and focal.

This communicates the series info with the ResolveCharm via the
inputted origin so the resulting origin will have the correct base
channel.

Fixes #822