Add ability to share juju model across integration tests
Currently the tests use the `CleanModel` context manager to initiate
and use a new model, by connecting to the current controller, adding a
brand new model, connecting to that model and returning it.

This is a prototype change that adds the `CachedModel` where we
initially do the same steps as the CleanModel, however, when we get
out of the context we don't destroy the model, we just clean it up for
the next test to use it. The purpose is to save up time not creating
and destorying models for every single test case.

Note that we have a bunch of tests that I believe this is not going to
work, but for those that will, all we need to do is to change the
model to the `CachedModel` and it should work as it is. For the other
tests that needs a fresh model to work, we can simply left them as
they are continue using the `CleanModel`.