Merge master to 3.1 (#845)
* [JUJU-3202] Add facades for 3.1.1. (#807)

* Add facades for 3.1.1.

* Update secrets backend test.

* Add destroy-units to destroy several units at once

Fixes #811

* Add integration test for destroy_units

* [JUJU-3517] Revisit _build_facades in connection (#826)

The main change here is that we no longer raise an exception in the
case of an unknown facade.

* Added 3.1.2 and 3.2-beta2 schemas.

* Add capability for deploying by revision

by passing the revision info through the ResolveCharm api call.

fixes #690

* Add example for deploy with revision

that deploys juju-qa-test --channel 2.0/stable --revision 22

* Add revision parameter in LocalDeployType.resolve

Note that --revision is only used for charmhub charms, so we're not
passing this info into the CharmOrigin for local charms.

* Add integration test for deploy by revision success

* Add validation for --revision flag to require --channel flag

This is needed because in libjuju we default to latest/stable if no
channel is specified. It is dangerous if we don't add this check
because without it the libjuju would deploy the wrong revision without
an exception if --revision is used but no --channel is given.

* Add validation for bundles to require either --revision or --channel

Otherwise fail early without making any additional API calls.

* Add integration test for making sure the --required and --channel
flags are validated before deploying charms or bundles

* [JUJU-3253] add missing force in bundle deployment (#815)

* Propagate force parameter to resolve function to enable force for remote bundles.

* Pass series info into origin for ResolveCharm

Without the series the ResolveCharm will select the latest base for
the charm. E.g. even if we want focal, the ResolveCharm will return
22.04 (jammy) for the base channel in the resulting origin if the
charm supports both jammy and focal.

This communicates the series info with the ResolveCharm via the
inputted origin so the resulting origin will have the correct base
channel.

Fixes #822

* Fix _resolve_charm errors

That are accidentally introduced in #825.

* _resolve_charm's force parameter is made optional
* fixed the _resolve_charm calls in bundle.py to have less number of
variables to unpack onto

* Change charm channel in bundle for test

* Rename example with consistent name

* Implement series selector for charm resolution

This selector will be used after the ResolveCharms api call, and use
the supported series and the base information returned to select the
appropriate series to construct the correct base for the charm. Then
it's given to the AddCharm in normal the deploy process.

* Add unit tests for series selector functionality

* Utilize the series selector to construct the correct base for charms

The main change is in the _resolve_charm that's used in the model and
bundle deploy code path. We pass in the series info whenever we can
into the ResolveCharm call, then use its result to run the series
selector, then construct the base using that and finally pass the
origin to the AddCharm. Note that the ResolveCharm API call is made no
more than once.

* Fix unit tests for bundle change runs

* Add integration test to challenge the resolver to find an old corrrect revision

This is basically testing what is manually reported by
@juanmanuel-tirado in the following review:

https://github.com/juju/python-libjuju/pull/830#issuecomment-1517536218

* [JUJU-3552] Prepare 3.1.2.1 release (#836)

* Add juju 3.1.2 missing facades. 
* Fix secrets-bakend-lifecycle test.
* Increase testing timeouts.
* Revisit postgresql charm to be downloaded.
---------

Co-authored-by: Juju bot <jujubot@users.noreply.github.com>
Co-authored-by: Caner Derici <caner.derici@canonical.com>

* Prepare release notes for 3.1.2.0. (#843)

* Prepare release notes for 3.1.2.0.

---------

Co-authored-by: Caner Derici <caner.derici@canonical.com>
Co-authored-by: Juju bot <jujubot@users.noreply.github.com>